split make_synthesis_engine
